pyproject.toml
[tool.poetry]
name = "discord-gift-exchange"
version = "0.1.0"
description = "discord bot to manage a christmas gift exchange, written in python"
authors = ["ryleu <riley.j.mclain@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
discord-py-interactions = "^4.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
